+++ packages/wagmi/src/utils/defaultWagmiReactConfig.ts
@@ -16,6 +16,7 @@ export type ConfigOptions = Partial<CreateConfigParameters> & {
     email?: boolean
     socials?: SocialProvider[]
     showWallets?: boolean
+    walletFeatures?: boolean
   }
   enableInjected?: boolean
   enableWalletConnect?: boolean
@@ -40,14 +41,12 @@ export function defaultWagmiConfig({
   ...wagmiConfig
 }: ConfigOptions): Config {
   const connectors: CreateConnectorFn[] = []
-  const transportsArr = chains.map(chain => [
-    chain.id,
-    getTransport({ chainId: chain.id, projectId })
-  ])
+  const transportsArr = chains.map(chain => [chain.id, getTransport({ chain, projectId })])
   const transports = Object.fromEntries(transportsArr)
   const defaultAuth = {
     email: true,
-    showWallets: true
+    showWallets: true,
+    walletFeatures: true
   }
 
   // Enabled by default
@@ -91,7 +90,8 @@ export function defaultWagmiConfig({
         options: { projectId },
         socials: mergedAuth.socials,
         email: mergedAuth.email,
-        showWallets: mergedAuth.showWallets
+        showWallets: mergedAuth.showWallets,
+        walletFeatures: mergedAuth.walletFeatures
       })
     )
   }