+++ apps/gallery/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @apps/gallery
 
+## 5.0.1
+
+### Patch Changes
+
+- fix: remove walletconnect restriction on names
+
+- Updated dependencies []:
+  - @web3modal/common@5.0.1
+  - @web3modal/ui@5.0.1
+
+## 5.0.0
+
+### Major Changes
+
+- Release V5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @web3modal/common@5.0.0
+  - @web3modal/ui@5.0.0
+
+## 5.0.0-cn-v5.0
+
+### Major Changes
+
+- Test V5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @web3modal/common@5.0.0-cn-v5.0
+  - @web3modal/ui@5.0.0-cn-v5.0
+
 ## 4.2.3
 
 ### Patch Changes