+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @web3modal/ui
 
+## 5.0.1
+
+### Patch Changes
+
+- fix: remove walletconnect restriction on names
+
+## 5.0.0
+
+### Major Changes
+
+- Release V5
+
+## 5.0.0-cn-v5.0
+
+### Major Changes
+
+- Test V5
+
 ## 4.2.3
 
 ### Patch Changes